Changed SearchResponse.Cluster objects to be immutable.
The SearchResponse.Clusters object now uses an immutable HashMap<String, AtomicReference<Cluster>>
rather than ConcurrentHashMap<String, Cluster>, since the concurrency control is
now a CAS operation on the underlying AtomicReference<Cluster>> not an overwrite into
the HashMap.

Changed took in SearchResponse.Cluster to be a TimeValue rather than long to match
the usage in SearchResponse.