Solved issue of why CCSearchIT was not getting skipped shards. It is now and it looks like I do not need to instrument the ShardSearch innerOnResponse callback - instead we can get the skipped list in the onListShards callback.
Added SearchTimeProvider as arg to notifyListShards and onListShards

Figured out that part of my problem before was that ccsMinimizeRoundtrips was
not being serialized in Clusters, so using that is problematic in IT tests
as it might flip to false if it goes through a Transport exchange.

Fixed that in this commit.